Agentic Video Summarizer/requirements.txt
streamlit
phi
google-generativeai
python-dotenv
duckduckgo-search
pathlib
tempfile